remove unused cursor code
